Move to a separate `example-tests` crate
