+++ Cargo.toml
@@ -48,19 +48,6 @@ serde_derive = "1.0"
 wasm-bindgen-test-crate-a = { path = 'tests/crates/a', version = '0.1' }
 wasm-bindgen-test-crate-b = { path = 'tests/crates/b', version = '0.1' }
 
-[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-anyhow = "1.0.58"
-futures-util = { version = "0.3.21", features = ["sink"] }
-hyper = { version = "0.14.20", features = ["server", "tcp", "http1"] }
-mozprofile = "0.8.0"
-mozrunner = "0.14.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1.20.0", features = ["macros", "time"] }
-tokio-tungstenite = "0.17.2"
-tower = { version = "0.4.13", features = ["make"] }
-tower-http = { version = "0.3.4", features = ["fs"] }
-
 [workspace]
 members = [
   "benchmarks",
@@ -68,6 +55,7 @@ members = [
   "crates/js-sys",
   "crates/test",
   "crates/test/sample",
+  "crates/example-tests",
   "crates/typescript-tests",
   "crates/web-sys",
   "crates/webidl",
