+++ crates/example-tests/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "example-tests"
+version = "0.1.0"
+authors = ["The wasm-bindgen Developers"]
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0.58"
+futures-util = { version = "0.3.21", features = ["sink"] }
+hyper = { version = "0.14.20", features = ["server", "tcp", "http1"] }
+mozprofile = "0.8.0"
+mozrunner = "0.14.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.20.0", features = ["macros", "time"] }
+tokio-tungstenite = "0.17.2"
+tower = { version = "0.4.13", features = ["make"] }
+tower-http = { version = "0.3.4", features = ["fs"] }
