+++ crates/example-tests/LICENSE-MIT
@@ -1,0 +1,1 @@
+../../LICENSE-MIT
