+++ crates/example-tests/README.md
@@ -1,0 +1,6 @@
+# example-tests
+
+Tests that none of our examples are broken, by opening them in a browser
+and checking that no errors get logged to the console.
+
+This currently only supports Firefox.
